[semver:minor] - CodeQL fixes & improvements
This fixes a bug where it was assumed a bin directory exists in the home directory of the user. This is the case for any of the provided CircleCI docker images, but not for most other ones.

We now always provide a configuration file for CodeQL. The default one is now part of this repository.